// web-frontend-Vite+React/src/api/models/Supplier.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Supplier = {
    createdAt?: string;
    updatedAt?: string;
    id?: string;
    name?: string;
    contactEmail?: string;
    phone?: string;
    address?: string;
};